Restructured code layout for modularity
