+++ requirements.txt
@@ -1,4 +1,12 @@
+certifi==2020.12.5
+chardet==4.0.0
 click==7.1.2
+docker==4.4.1
 falcon==2.0.0
-psutil==5.7.3
+idna==2.10
+psutil==5.8.0
+requests==2.25.1
+six==1.15.0
+urllib3==1.26.3
+websocket-client==0.57.0
 Werkzeug==1.0.1
